backend/requirements.txt
annotated-types==0.6.0
anyio @ file:///private/tmp/nix-build-python3.10-anyio-3.6.2.drv-0/source/dist/anyio-3.6.2-py3-none-any.whl
asgiref==3.8.1
certifi @ file:///private/tmp/nix-build-python3.10-certifi-2022.12.07.drv-0/source/dist/certifi-2022.12.7-py3-none-any.whl
click @ file:///private/tmp/nix-build-python3.10-click-8.1.3.drv-0/click-8.1.3/dist/click-8.1.3-py3-none-any.whl
distlib @ file:///private/tmp/nix-build-python3.10-distlib-0.3.6.drv-0/distlib-0.3.6/dist/distlib-0.3.6-py2.py3-none-any.whl
Django==5.0.4
django-extensions==3.2.3
django-ninja==1.1.0
django-types==0.19.1
docutils @ file:///private/tmp/nix-build-python3.10-docutils-0.19.drv-0/docutils-0.19/dist/docutils-0.19-py3-none-any.whl
editables @ file:///private/tmp/nix-build-python3.10-editables-0.3.drv-0/editables-0.3/dist/editables-0.3-py3-none-any.whl
filelock @ file:///private/tmp/nix-build-python3.10-filelock-3.9.0.drv-0/filelock-3.9.0/dist/filelock-3.9.0-py3-none-any.whl
h11 @ file:///private/tmp/nix-build-python3.10-h11-0.14.0.drv-0/h11-0.14.0/dist/h11-0.14.0-py3-none-any.whl
hatch @ file:///private/tmp/nix-build-hatch-1.6.3.drv-0/hatch-1.6.3/dist/hatch-1.6.3-py3-none-any.whl
hatchling @ file:///private/tmp/nix-build-python3.10-hatchling-1.13.0.drv-0/hatchling-1.13.0/dist/hatchling-1.13.0-py3-none-any.whl
httpcore @ file:///private/tmp/nix-build-python3.10-httpcore-0.16.3.drv-0/source/dist/httpcore-0.16.3-py3-none-any.whl
httpx @ file:///private/tmp/nix-build-python3.10-httpx-0.23.3.drv-0/source/dist/httpx-0.23.3-py3-none-any.whl
hyperlink @ file:///private/tmp/nix-build-python3.10-hyperlink-21.0.0.drv-0/hyperlink-21.0.0/dist/hyperlink-21.0.0-py2.py3-none-any.whl
idna @ file:///private/tmp/nix-build-python3.10-idna-3.4.drv-0/idna-3.4/dist/idna-3.4-py3-none-any.whl
importlib-metadata @ file:///private/tmp/nix-build-python3.10-importlib-metadata-6.0.0.drv-0/importlib_metadata-6.0.0/dist/importlib_metadata-6.0.0-py3-none-any.whl
jaraco.classes @ file:///private/tmp/nix-build-python3.10-jaraco.classes-3.1.1.drv-0/source/dist/jaraco.classes-3.1.1-py3-none-any.whl
keyring @ file:///private/tmp/nix-build-python3.10-keyring-23.13.1.drv-0/keyring-23.13.1/dist/keyring-23.13.1-py3-none-any.whl
markdown-it-py @ file:///private/tmp/nix-build-python3.10-markdown-it-py-2.2.0.drv-0/source/dist/markdown_it_py-2.2.0-py3-none-any.whl
mdurl @ file:///private/tmp/nix-build-python3.10-mdurl-0.1.2.drv-0/source/dist/mdurl-0.1.2-py3-none-any.whl
more-itertools @ file:///private/tmp/nix-build-python3.10-more-itertools-9.1.0.drv-0/more-itertools-9.1.0/dist/more_itertools-9.1.0-py3-none-any.whl
orjson==3.10.1
packaging @ file:///private/tmp/nix-build-python3.10-packaging-23.0.drv-0/packaging-23.0/dist/packaging-23.0-py3-none-any.whl
pathspec @ file:///private/tmp/nix-build-python3.10-pathspec-0.11.0.drv-0/pathspec-0.11.0/dist/pathspec-0.11.0-py3-none-any.whl
pexpect @ file:///private/tmp/nix-build-python3.10-pexpect-4.8.0.drv-0/pexpect-4.8.0/dist/pexpect-4.8.0-py2.py3-none-any.whl
platformdirs @ file:///private/tmp/nix-build-python3.10-platformdirs-3.0.0.drv-0/source/dist/platformdirs-3.0.0-py3-none-any.whl
pluggy @ file:///private/tmp/nix-build-python3.10-pluggy-1.0.0.drv-0/pluggy-1.0.0/dist/pluggy-1.0.0-py2.py3-none-any.whl
poetry @ file:///private/tmp/nix-build-python3.10-poetry-1.4.2.drv-0/source/dist/poetry-1.4.2-py3-none-any.whl
ptyprocess @ file:///private/tmp/nix-build-python3.10-ptyprocess-0.7.0.drv-0/ptyprocess-0.7.0/dist/ptyprocess-0.7.0-py3-none-any.whl
pydantic==2.7.1
pydantic_core==2.18.2
Pygments @ file:///private/tmp/nix-build-python3.10-pygments-2.14.0.drv-0/Pygments-2.14.0/dist/Pygments-2.14.0-py3-none-any.whl
pyperclip @ file:///private/tmp/nix-build-python3.10-pyperclip-1.8.2.drv-0/pyperclip-1.8.2/dist/pyperclip-1.8.2-py3-none-any.whl
rfc3986 @ file:///private/tmp/nix-build-python3.10-rfc3986-2.0.0.drv-0/rfc3986-2.0.0/dist/rfc3986-2.0.0-py2.py3-none-any.whl
rich @ file:///private/tmp/nix-build-python3.10-rich-13.3.1.drv-0/source/dist/rich-13.3.1-py3-none-any.whl
shellingham @ file:///private/tmp/nix-build-python3.10-shellingham-1.5.0.post1.drv-0/source/dist/shellingham-1.5.0.post1-py2.py3-none-any.whl
six @ file:///private/tmp/nix-build-python3.10-six-1.16.0.drv-0/six-1.16.0/dist/six-1.16.0-py2.py3-none-any.whl
sniffio @ file:///private/tmp/nix-build-python3.10-sniffio-1.3.0.drv-0/sniffio-1.3.0/dist/sniffio-1.3.0-py3-none-any.whl
sqlparse==0.5.0
toml @ file:///private/tmp/nix-build-python3.10-toml-0.10.2.drv-0/toml-0.10.2/dist/toml-0.10.2-py2.py3-none-any.whl
tomli @ file:///private/tmp/nix-build-python3.10-tomli-2.0.1.drv-0/source/dist/tomli-2.0.1-py3-none-any.whl
tomli_w @ file:///private/tmp/nix-build-python3.10-tomli-w-1.0.0.drv-0/source/dist/tomli_w-1.0.0-py3-none-any.whl
tomlkit @ file:///private/tmp/nix-build-python3.10-tomlkit-0.11.6.drv-0/tomlkit-0.11.6/dist/tomlkit-0.11.6-py3-none-any.whl
types-psycopg2==2.9.21.20240417
typing_extensions==4.11.0
userpath @ file:///private/tmp/nix-build-python3.10-userpath-1.8.0.drv-0/userpath-1.8.0/dist/userpath-1.8.0-py3-none-any.whl
virtualenv @ file:///private/tmp/nix-build-python3.10-virtualenv-20.19.0.drv-0/virtualenv-20.19.0/dist/virtualenv-20.19.0-py3-none-any.whl
zipp @ file:///private/tmp/nix-build-python3.10-zipp-3.15.0.drv-0/zipp-3.15.0/dist/zipp-3.15.0-py3-none-any.whl
